Move statefull components to containers directory
